src/kinematics/requirements.txt
ansitable==0.11.4
cfgv==3.4.0
colored==2.3.0
contourpy==1.3.1
cycler==0.12.1
distlib==0.3.9
filelock==3.17.0
fonttools==4.56.0
identify==2.6.7
kiwisolver==1.4.8
matplotlib==3.10.0
nodeenv==1.9.1
numpy==1.26.4
packaging==24.2
pgraph-python==0.6.3
pillow==11.1.0
platformdirs==4.3.6
pre_commit==4.1.0
progress==1.6
pyparsing==3.2.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
roboticstoolbox-python==1.1.1
rtb-data==1.0.1
scipy==1.15.1
six==1.17.0
spatialgeometry==1.1.0
spatialmath-python==1.1.14
swift-sim==1.1.0
typing_extensions==4.12.2
virtualenv==20.29.2
websockets==14.2
